+++ src/main/java/nl/knaw/dans/catalog/DdVaultCatalogApplication.java
@@ -29,8 +29,6 @@ import io.dropwizard.views.ViewBundle;
 import nl.knaw.dans.catalog.cli.ReindexCommand;
 import nl.knaw.dans.catalog.core.SearchIndex;
 import nl.knaw.dans.catalog.core.UseCases;
-import nl.knaw.dans.catalog.core.mappers.OcflObjectVersionMapper;
-import nl.knaw.dans.catalog.core.mappers.TarMapper;
 import nl.knaw.dans.catalog.core.solr.OcflObjectMetadataReader;
 import nl.knaw.dans.catalog.core.solr.SolrIndex;
 import nl.knaw.dans.catalog.db.OcflObjectVersionDao;
@@ -39,8 +37,7 @@ import nl.knaw.dans.catalog.resources.ArchiveDetailResource;
 import nl.knaw.dans.catalog.resources.DefaultApiResource;
 import nl.knaw.dans.catalog.resources.ErrorView;
 import nl.knaw.dans.catalog.resources.OcflObjectApiResource;
-import nl.knaw.dans.catalog.resources.TarAPIResource;
-import org.mapstruct.factory.Mappers;
+import nl.knaw.dans.catalog.resources.TarApiResource;
 
 import javax.ws.rs.core.MediaType;
 
@@ -70,7 +67,7 @@ public class DdVaultCatalogApplication extends Application<DdVaultCatalogConfigu
         var useCases = buildUseCases(configuration);
 
         environment.jersey().register(new DefaultApiResource());
-        environment.jersey().register(new TarAPIResource(useCases));
+        environment.jersey().register(new TarApiResource(useCases));
         environment.jersey().register(new OcflObjectApiResource(useCases));
         environment.jersey().register(new ArchiveDetailResource(useCases));
         environment.jersey().register(new ErrorEntityWriter<ErrorMessage, View>(MediaType.TEXT_HTML_TYPE, View.class) {
@@ -87,8 +84,6 @@ public class DdVaultCatalogApplication extends Application<DdVaultCatalogConfigu
         var searchIndex = new SolrIndex(configuration.getSolr(), ocflObjectMetadataReader);
         var ocflObjectVersionDao = new OcflObjectVersionDao(hibernateBundle.getSessionFactory());
         var tarDao = new TarDao(hibernateBundle.getSessionFactory());
-        var ocflObjectVersionMapper = Mappers.getMapper(OcflObjectVersionMapper.class);
-        var tarMapper = Mappers.getMapper(TarMapper.class);
 
         return new UnitOfWorkAwareProxyFactory(hibernateBundle)
             .create(UseCases.class,
