+++ src/main/java/nl/knaw/dans/catalog/cli/ReindexCommand.java
@@ -24,13 +24,10 @@ import net.sourceforge.argparse4j.inf.Namespace;
 import nl.knaw.dans.catalog.DdVaultCatalogConfiguration;
 import nl.knaw.dans.catalog.core.SearchIndex;
 import nl.knaw.dans.catalog.core.UseCases;
-import nl.knaw.dans.catalog.core.mappers.OcflObjectVersionMapper;
-import nl.knaw.dans.catalog.core.mappers.TarMapper;
 import nl.knaw.dans.catalog.core.solr.OcflObjectMetadataReader;
 import nl.knaw.dans.catalog.core.solr.SolrIndex;
 import nl.knaw.dans.catalog.db.OcflObjectVersionDao;
 import nl.knaw.dans.catalog.db.TarDao;
-import org.mapstruct.factory.Mappers;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -63,23 +60,17 @@ public class ReindexCommand extends ConfiguredCommand<DdVaultCatalogConfiguratio
         var searchIndex = new SolrIndex(configuration.getSolr(), ocflObjectMetadataReader);
         var ocflObjectVersionDao = new OcflObjectVersionDao(hibernateBundle.getSessionFactory());
         var tarDao = new TarDao(hibernateBundle.getSessionFactory());
-        var ocflObjectVersionMapper = Mappers.getMapper(OcflObjectVersionMapper.class);
-        var tarMapper = Mappers.getMapper(TarMapper.class);
 
         var useCases = new UnitOfWorkAwareProxyFactory(hibernateBundle)
             .create(UseCases.class,
                 new Class[] {
                     OcflObjectVersionDao.class,
                     TarDao.class,
-                    OcflObjectVersionMapper.class,
-                    TarMapper.class,
                     SearchIndex.class,
                 },
                 new Object[] {
                     ocflObjectVersionDao,
                     tarDao,
-                    ocflObjectVersionMapper,
-                    tarMapper,
                     searchIndex
                 }
             );
