+++ src/test/java/nl/knaw/dans/catalog/resources/ArchiveDetailResourceTest.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package nl.knaw.dans.catalog.resources.web;
+package nl.knaw.dans.catalog.resources;
 
 import com.codahale.metrics.MetricRegistry;
 import freemarker.template.Configuration;
